refa: fazendo o componente de verificacao de token e retirando os hardcodados da aplicacao

+++ tupan/src/app/_api/delete/parametro.ts
@@ -1,11 +1,22 @@
 import { api_route } from '..';
 
+// Função para obter o token armazenado
+const obterToken = (): string | null => {
+  return localStorage.getItem('token');
+};
+
 export const deletarParametro = async (id: number): Promise<void> => {
+  const token = obterToken(); // Obtendo o token armazenado
+
+  if (!token) {
+    throw new Error('Token não encontrado. Por favor, faça login.');
+  }
+
   try {
     const response = await fetch(`${api_route}/parametros/${id}`, {
       method: 'DELETE',
       headers: {
-        'Authorization': `Token 1112a98d58500b7a165191fc56b2a9c1513413e8`,
+        'Authorization': `Token ${token}`, // Usando o token dinamicamente
         'Content-Type': 'application/json',
       },
     });
